Product Page v2
IMPORTANT PART OF THIS: The "quantity" field of the shopping cart is accurate. So, it can be used for things like calculating subtotal for each item.

Add to cart buttons actually add the object to localStorage now, under the "shoppingCart" key.

Shopping cart's quantity value is now accurate. If an item isn't in the cart, its quantity is set to 1. If it is, it's not added to the array (since it's already there) and the existing version of it is increased by 1.